Ensure options in users list modal work well
