+++ packages/smooth_app/lib/pages/history_page.dart
@@ -3,7 +3,7 @@ import 'package:provider/provider.dart';
 import 'package:smooth_app/data_models/product_list.dart';
 import 'package:smooth_app/database/dao_product_list.dart';
 import 'package:smooth_app/database/local_database.dart';
-import 'package:smooth_app/pages/product/common/product_list_modal.dart';
+import 'package:smooth_app/pages/product/common/product_list_page.dart';
 
 class HistoryPage extends StatefulWidget {
   const HistoryPage();
